+++ apps/blog/src/app/stores/index.ts
@@ -1,1 +1,2 @@
-export { StoreProvider, useStore } from "./store-context";
+export { StoreProvider } from "./store-context";
+export { useStore } from "./use";
