+++ apps/blog/src/app/stores/root-store.ts
@@ -1,5 +1,3 @@
-import { makeAutoObservable } from "mobx";
-
 import { EditorStore } from "./editor";
 import { ViewStore } from "./viewstore";
 
@@ -10,6 +8,5 @@ export class RootStore {
   constructor() {
     this.editor = new EditorStore();
     this.view = new ViewStore();
-    makeAutoObservable(this, {}, { autoBind: true });
   }
 }
