+++ apps/blog/src/app/stores/store-context.tsx
@@ -2,11 +2,11 @@
 
 import type { ReactNode } from "react";
 import type { Optional } from "ts-roids";
-import { createContext, useContext, useMemo } from "react";
+import { createContext, useMemo } from "react";
 
 import { RootStore } from "./root-store";
 
-const StoreContext = createContext<Optional<RootStore>>(null);
+export const StoreContext = createContext<Optional<RootStore>>(null);
 
 interface StoreProviderProps {
   children: ReactNode;
@@ -19,12 +19,4 @@ export function StoreProvider({ children }: StoreProviderProps) {
   return (
     <StoreContext.Provider value={store}>{children}</StoreContext.Provider>
   );
-}
-
-export function useStore(): { store: RootStore } {
-  const store = useContext(StoreContext);
-  if (!store) {
-    throw new Error("useStore must be used within StoreProvider");
-  }
-  return { store };
 }
